+++ packages/core/src/components/skeleton/skeleton.mdx
@@ -2,6 +2,8 @@
 title: Skeleton
 ---
 
+import { skeleton } from "./skeleton_css";
+
 # Skeleton
 
 Skeletons provide a loading state that mimics the shape of your yet-to-load content.
@@ -11,7 +13,7 @@ an element through `className`, will cover its content with a loading animation.
 The skeleton inherits the dimensions of whatever element the class is applied
 to, so you should supply a placeholder while awaiting real content.
 
-<CssExample value="skeleton" />
+<CssExample data={skeleton} />
 
 ## CSS
